Isolate inplace test
Turning on inplace_mode for building experiments monkey patches the slicedimage `Tile` class.  This is regarded as acceptable, as experiment builders are all isolated scripts.

However, in the test environment, the monkey patch persists, and can affect subsequent tests.  Therefore, if `starfish/experiment/builder/test/test_inplace.py` was run immediately before `starfish/imagestack/test/test_imagestack.py`, `test_imagestack_export` would spuriously fail.

This PR isolates the inplace test into its own subprocess, much like it would likely be run in real life.  It also adds a note to the documentation that in-place mode should be done in isolated scripts.

Test plan: `pytest starfish/experiment/builder/test/test_inplace.py starfish/imagestack/test/test_imagestack.py` passes now.
